sorteio em funcionamento

+++ jogo/websocket_triggers_bilhete.py
@@ -9,10 +9,11 @@ def event_bilhete_sorteio(id_sorteio):
     termino do sorteio
     '''
     layer = get_channel_layer()
-    async_to_sync(layer.group_send)('sorteio-'+str(id_sorteio), {
-        'type': 'events.bilhete.sorteio',
-        'id_sorteio':id_sorteio
-    })
+    if layer:
+        async_to_sync(layer.group_send)('sorteio-'+str(id_sorteio), {
+            'type': 'events.bilhete.sorteio',
+            'id_sorteio':id_sorteio
+        })
 
 
 def event_fechar_grupo_sorteio(id_sorteio):
@@ -20,17 +21,17 @@ def event_fechar_grupo_sorteio(id_sorteio):
     termino do sorteio
     '''
     layer = get_channel_layer()
+    if layer:
+        grupo = GrupoWebSocket.objects.filter(partida__id=id_sorteio).first()
+        if grupo:
 
-    grupo = GrupoWebSocket.objects.filter(partida__id=id_sorteio).first()
-    if grupo:
+            canais = grupo.grupocanal_set.all()
+            for canal in canais:
+                async_to_sync(layer.group_discard)(grupo.nome, canal.nome)
 
-        canais = grupo.grupocanal_set.all()
-        for canal in canais:
-            async_to_sync(layer.group_discard)(grupo.nome, canal.nome)
-
-        print(grupo.nome)
-        async_to_sync(layer.group_send)(grupo.nome, {"type":"websocket_disconnect"})
-        grupo.delete()
+            print(grupo.nome)
+            async_to_sync(layer.group_send)(grupo.nome, {"type":"websocket_disconnect"})
+            grupo.delete()
 
 
 def event_bilhete_partida(id_sorteio):
@@ -38,13 +39,14 @@ def event_bilhete_partida(id_sorteio):
         termino do sorteio
         '''
     layer = get_channel_layer()
-    agora = datetime.now()
-    sorteio_ids = [partida.id for partida in Partida.objects.filter(data_partida__gt=agora)]
-    if not sorteio_ids:
-        sorteio_ids = [id_sorteio]
-    for id in sorteio_ids:
-        async_to_sync(layer.group_send)('sorteio-' + str(id), {
-            'type': 'events.bilhete.partida',
-            'id_sorteio': id_sorteio
-        })
+    if layer:
+        agora = datetime.now()
+        sorteio_ids = [partida.id for partida in Partida.objects.filter(data_partida__gt=agora)]
+        if not sorteio_ids:
+            sorteio_ids = [id_sorteio]
+        for id in sorteio_ids:
+            async_to_sync(layer.group_send)('sorteio-' + str(id), {
+                'type': 'events.bilhete.partida',
+                'id_sorteio': id_sorteio
+            })
